LogFile Creation System
This is the system to create the Log File, now need to work on the physical logging system. The createFile and Create directory files should always be ran when the program starts. Just in case a new change to pc memory occurs. 

Also allows the logfile to reset itself when the new year comes along. It runs on the system memory (clock) currently... 

Need to work on the backup system as well